gather relevant code in plugin module
